+++ R/req-cache.R
@@ -132,6 +132,10 @@ cache_post_fetch <- function(req, resp, path = NULL) {
 
     # Replace body with cached result
     resp$body <- cache_body(req, path)
+
+    # Combine headers
+    resp$headers <- cache_headers(req, resp)
+
     resp
   } else if (resp_is_cacheable(resp)) {
     if (debug) cli::cli_text("Saving response to cache {.val {hash(req$url)}}")
@@ -144,6 +148,7 @@ cache_post_fetch <- function(req, resp, path = NULL) {
 
 cache_body <- function(req, path = NULL) {
   body <- cache_get(req)$body
+
   if (is.null(path)) {
     return(body)
   }
@@ -154,6 +159,12 @@ cache_body <- function(req, path = NULL) {
     writeBin(body, path)
   }
   new_path(path)
+}
+
+cache_headers <- function(req, resp) {
+  # https://www.rfc-editor.org/rfc/rfc7232#section-4.1
+  cached_headers <- cache_get(req)$headers
+  as_headers(modify_list(cached_headers, !!!resp$headers))
 }
 
 # Caching headers ---------------------------------------------------------